% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automobile.R
\docType{data}
\name{automobile}
\alias{automobile}
\title{Automobiles data}
\format{
A dataset with 205 observations and 26 attributes.
\describe{
  \item{symboling}{-3, -2, -1, 0, 1, 2, 3.}
  \item{normalized-losses}{continuous from 65 to 256.}
  \item{make}{alfa-romero, audi, bmw, chevrolet, dodge, honda,isuzu, jaguar, mazda, mercedes-benz, mercury,mitsubishi, nissan, peugot, plymouth, porsche,renault, saab, subaru, toyota, volkswagen, volvo}
  \item{fuel-type}{diesel, gas.}
  \item{aspiration}{std, turbo.}
  \item{num-of-doors}{four, two.}
  \item{body-style}{hardtop, wagon, sedan, hatchback, convertible.}
  \item{drive-wheels}{4wd, fwd, rwd.}
  \item{engine-location}{front, rear.}
  \item{wheel-base}{continuous from 86.6 120.9.}
  \item{length}{continuous from 141.1 to 208.1.}
  \item{width}{continuous from 60.3 to 72.3.}
  \item{height}{continuous from 47.8 to 59.8.}
  \item{curb-weight}{continuous from 1488 to 4066.}
  \item{engine-type}{dohc, dohcv, l, ohc, ohcf, ohcv, rotor.}
  \item{num-of-cylinders}{eight, five, four, six, three, twelve, two.}
  \item{engine-size}{continuous from 61 to 326.}
  \item{fuel-system}{1bbl, 2bbl, 4bbl, idi, mfi, mpfi, spdi, spfi.}
  \item{bore}{continuous from 2.54 to 3.94.}
  \item{stroke}{continuous from 2.07 to 4.17.}
  \item{compression-ratio}{continuous from 7 to 23.}
  \item{horsepower}{continuous from 48 to 288.}
  \item{peak-rpm}{continuous from 4150 to 6600.}
  \item{city-mpg}{continuous from 13 to 49.}
  \item{highway-mpg}{continuous from 16 to 54.}
  \item{price}{continuous from 5118 to 45400.}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/automobile}
}
\usage{
data(automobile)
}
\description{
This dataset contains the details about automobiles from 1985 Ward's automotive yearbook.
}
\keyword{datasets}
