% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R
\docType{data}
\name{pdb}
\alias{pdb}
\title{Planning Database (Published in year 2015)}
\format{
A dataset with 815 observations and 344 attributes.
}
\source{
\url{https://www.census.gov/data/datasets/2015/adrm/research/2015-planning-database.html}
}
\usage{
data(pdb)
}
\description{
The Planning Database (pdb) contains selected data from the 2010 Census and the 2009-2013, 5-years American Community Survey (ACS) estimates.
}
\keyword{datasets}
