% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raman.R
\docType{data}
\name{Raman}
\alias{Raman}
\title{Raman spectroscopy}
\format{
An object of class \code{data.frame} with 69 rows and 1100 columns.
}
\usage{
data(raman)
}
\description{
The Raman dataset contains 69 samples of fatty acid information in terms of 
percentage of total sample weight and percentage of total fat content
}
\references{
Naes T., Tomic O., Afseth N.K., Segtnan V., Mage I. (2013)
Multi-block regression based on combinations of orthogonalisation, pls-regression and canonical
correlation analysis. Chemom. Intell. Lab. Syst. 124, 32-42.
}
\keyword{datasets}
