% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB.R
\docType{data}
\name{PDB}
\alias{PDB}
\title{Planning database published in year 2015.}
\format{
A dataset with 816 observations and 344 attributes.
}
\source{
\url{https://www.census.gov/data/datasets/2015/adrm/research/2015-planning-database.html}
}
\usage{
data(PDB)
}
\description{
Planning Database (PDB) contains selected 2010 Census and selected 2009-2013, 5-years American Community Survey (ACS) estimates.
}
\keyword{datasets}
