% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostate.R
\docType{data}
\name{prostate}
\alias{prostate}
\title{Prostate Levels}
\format{
An object of class \code{data.frame} with 97 rows and 9 columns.
}
\usage{
data(prostate)
}
\description{
The dataset comprises the level of prostate-specific antigen
associated with eight clinical measures in 97 male patients
undergoing a radical prostatectomy.
}
\references{
Stamey, T. A., Kabalin, J. N., McNeal, J. E., Johnstone, I. M., Freiha, F., Redwine, E. A. et al.
(1989). Prostate specific Antigen in the Diagnosis and Treatment of Adenocarcinoma of the
Prostate. II. \emph{Radical prostatectomy treated patients. The Journal of Urology}.
141, 1076, 1083.
}
\keyword{datasets}
