% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raman.R
\docType{data}
\name{raman}
\alias{raman}
\title{Raman Spectroscopy}
\format{
An object of class \code{data.frame} with 69 rows and 1100 columns.
}
\usage{
data(raman)
}
\description{
The Raman dataset contains 69 samples of providing fatty acid information in terms of
the percentage of total sample weight and the percentage of total fat content
}
\references{
Naes T., Tomic O., Afseth N.K., Segtnan V., Mage I. (2013)
Multi-Block Regression Based on Combinations of Orthogonalisation, Pls-Regression and Canonical
Correlation Analysis. \emph{Chemometrics and Intelligent Laboratory Systems}. 124, 32-42.
}
\keyword{datasets}
