% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recumbent.R
\docType{data}
\name{Recumbent}
\alias{Recumbent}
\title{Recumbent cows}
\format{
A dataset with 9 columns and 435 rows.
}
\source{
Clark, R. G., Henderson, H. V., Hoggard, G. K. Ellison, R. S. and Young, B. J. (1987). 
The abiltiy of biochemical and haematolgical tests to predict recovery in periparturient 
recumbent cows. NZ Veterinary Journal, 35, 126-133.
}
\usage{
data(Recumbent)
}
\description{
For unknown reasons, pregnant dairy cows can become
recumbent-they lie down-either shortly before or after calving. This condition
can be serious, and frequently leads to death of the cow. Clark, Henderson,
Hoggard, Ellison and Young (1987) analyze data collected at the Ruakura (N.Z.)
Animal Health Laboratory on a sample of recumbent cows.
}
\keyword{datasets}
